+++ src/view/OutputView.js
@@ -1,6 +1,6 @@
 const { Console } = require('@woowacourse/mission-utils');
-const { GAME_MESSAGES } = require('./Messages');
-const { BRIDGE_CONSTANTS } = require('./GameConstants');
+const { GAME_MESSAGES } = require('../constant/Messages');
+const { BRIDGE_CONSTANTS } = require('../constant/GameConstants');
 
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
